+++ src/Ares/Rule/RegexRule.php
@@ -3,51 +3,49 @@
 declare(strict_types=1);
 
 /**
- * TypeRule.php
+ * RegexRule.php
  *
  * @author Michael Leßnau <michael.lessnau@gmail.com>
- * @since  2019-03-20
+ * @since  2019-03-30
  */
 
 namespace Ares\Rule;
 
 use Ares\Context;
+use Ares\Exception\InapplicableValidationRuleException;
 use Ares\Exception\InvalidValidationRuleArgsException;
+use Ares\Rule\TypeRule;
 use Ares\Schema\Type;
 
 /**
- * Class TypeRule
+ * Class RegexRule
  */
-class TypeRule implements RuleInterface
+class RegexRule implements RuleInterface
 {
-    const ID            = 'type';
-    const ERROR_MESSAGE = 'Invalid type';
-
-    /* @const array TYPE_MAPPING maps PHP types to validator specific types */
-    const TYPE_MAPPING = [
-        'array'   => [Type::LIST, Type::MAP],
-        'boolean' => [Type::BOOLEAN],
-        'double'  => [Type::FLOAT],
-        'integer' => [Type::INTEGER],
-        'string'  => [Type::STRING],
-    ];
+    const ID            = 'regex';
+    const ERROR_MESSAGE = 'Value invalid';
 
     /**
      * @param mixed         $args    Validation rule configuration.
      * @param mixed         $data    Input data.
      * @param \Ares\Context $context Validation context.
      * @return boolean
+     * @throws \Ares\Exception\InapplicableValidationRuleException
      * @throws \Ares\Exception\InvalidValidationRuleArgsException
      */
     public function validate($args, $data, Context $context): bool
     {
-        if (!in_array($args, Type::getValues(), true)) {
+        $schema = $context->getSchema();
+
+        if ($schema[TypeRule::ID] !== Type::STRING) {
+            throw new InapplicableValidationRuleException('This rule applies to <string> types only');
+        }
+
+        if (!is_string($args)) {
             throw new InvalidValidationRuleArgsException('Invalid args: ' . json_encode($args));
         }
 
-        $phpType = gettype($data);
-
-        if (isset(self::TYPE_MAPPING[$phpType]) && in_array($args, self::TYPE_MAPPING[$phpType], true) || $data === null) {
+        if (preg_match($args, $data) === 1) {
             return true;
         }