+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+
+## [v0.2.0] 2019-03-31
+### Added
+- ```email``` validation rule (string)
+- ```url``` validation rule (string)
+- ```min``` validation rule (float, integer)
+- ```max``` validation rule (float, integer)
+- ```regex``` validation rule (string)
+- ```list``` type
+
 ## [v0.1.0] 2019-03-28
 ### Added
 - CHANGELOG.md to track changes between releases of the library